moved files around, redid the layout of the theme modal ... again
