+++ app/lish/[did]/[publication]/dashboard/page.tsx
@@ -66,10 +66,10 @@ export default async function Publication(props: {
 
   let record = publication?.record as PubLeafletPublication.Record | null;
   let backgroundAlpha =
-    PubLeafletThemeColor.isRgba(record?.theme?.backgroundColor) &&
-    record?.theme?.backgroundColor?.a;
+    PubLeafletThemeColor.isRgba(record?.theme?.pageBackground) &&
+    record?.theme?.pageBackground?.a;
 
-  let hasBackground = !!record?.theme?.backgroundImage && backgroundAlpha !== 0;
+  let hasBackground = !!record?.theme?.pageBackground && backgroundAlpha !== 0;
 
   if (!publication || identity.atp_did !== publication.identity_did)
     return <PubNotFound />;
