+++ components/Toggle.tsx
@@ -3,18 +3,20 @@ import { theme } from "tailwind.config";
 export const Toggle = (props: {
   toggleOn: boolean;
   setToggleOn: (s: boolean) => void;
+  disabledColor1?: string;
+  disabledColor2?: string;
 }) => {
   return (
     <button
-      className="toggle selected-outline transparent-outline flex items-center h-[20px] w-7 bg-test rounded-md border-border"
+      className="toggle selected-outline transparent-outline flex items-center h-[20px] w-6 rounded-md border-border"
       style={{
         border: props.toggleOn
           ? "1px solid " + theme.colors["accent-2"]
-          : "1px solid " + theme.colors["border-light"],
+          : "1px solid " + props.disabledColor2 || theme.colors["border-light"],
         justifyContent: props.toggleOn ? "flex-end" : "flex-start",
         background: props.toggleOn
           ? theme.colors["accent-1"]
-          : theme.colors["tertiary"],
+          : props.disabledColor1 || theme.colors["tertiary"],
       }}
       onClick={() => props.setToggleOn(!props.toggleOn)}
     >
@@ -23,7 +25,7 @@ export const Toggle = (props: {
         style={{
           background: props.toggleOn
             ? theme.colors["accent-2"]
-            : theme.colors["border-light"],
+            : props.disabledColor2 || theme.colors["border-light"],
         }}
       />
     </button>
